+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## v0.1.1, 02/18/2014
 
 - Add: Search for movies playing near you
-- Add: Search for movie titles
+- Add: Search by title
+- Add: Search by movie title
 - Add: setup.py script
 - Add: pytest for testing purposes
+- Add: Search by tv title
+- Add: Search by videogame title